Add files via upload
Lesson 1. A console application that will ask for the current city and display: "It's hot in N city right now", where N is the city entered from the keyboard. If the temperature is from -50 to 15, then it's cold, from 15 to 25 is normal and from 25 to 50 is hot. Otherwise, it's a disaster.